requirements.txt
Django==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.7.0
